trip-planner/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
azure-ai-agents==1.2.0b6
azure-ai-projects==1.0.0
azure-core==1.38.2
azure-identity==1.25.2
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
cloudevents==1.12.0
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
fastapi==0.135.1
frozenlist==1.8.0
google-crc32c==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
MarkupSafe==3.0.3
mcp==1.26.0
more-itertools==10.8.0
msal==1.35.1
msal-extensions==1.3.1
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.4.2
openai==2.26.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
parse==1.21.1
pathable==0.4.4
prance==25.4.8.0
propcache==0.4.1
pybars4==0.9.13
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.13.1
pydantic_core==2.33.2
pyee==13.0.1
PyJWT==2.11.0
pylibsrtp==1.0.0
PyMeta3==0.5.1
pyOpenSSL==25.3.0
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.30.0
ruamel.yaml==0.19.1
scipy==1.17.1
semantic-kernel==1.40.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.2
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.23.0
zipp==3.23.0
